# tracer: limit the scope of setting the priority when auto dropped