@katiehockman 2023-08-03
I'm going to hold off on merging this until @ajgajg1134 is able to look at this, since he made the original change. I want to make sure I'm not missing something here, and the performance impact is ~negligible anyway, across an entire trace.